Initial implementation of FSL Plugin for Celeritas offload
Implement a FSL UserAction plugin that offloads all EM particles
to Celeritas. Use of Celeritas via the SimpleOffload mechanism is
taken largely as-is from the atlas-tilecal-integration project:

- https://github.com/celeritas-project/atlas-tilecal-integration

The primary modification is disabling use of SDs pending implementation
of plugin(s) for these.

Provide example FSL configuration file and TestEM3 GDML file for
quick and dirty testing.