Explicitly link to Celeritas::accel_final
Building and running the Celeritas plugin on CUDA/GPU revealing
an undefined reference error on plugin loading. This was traced
to `celeritas_target_link_libraries` not adding the `_final`
targets to the plugin link. It only does this for executables.

Workaround by using pure target_link_libraries and the explicit
list of targets needed, allowing for pure CPU and CPU/GPU builds.