Use Acceleritas flag to create G4Track when available
Celeritas PR 839 introduced a flag in Acceleritas SDSetupOptions to control
reconstitution of G4Tracks in the G4Steps passed back to host-side SDs. It
is off by default, leading to segfaults in thr ATLTileCal SDs, which use the
track info.

Introduce CMake-time compile check for presence of SDSetupOptions track flag.
When available, always set to true so that track info is available in the G4Step.